Fix all style files due to overwrite bug.
Before the style files relied on a quirk in the LESS compiler in order to overwrite the variables
of the default style with their custom colors. This no longer works, so now the styles have
been seperated into variables and rules files, so the themes can import the correct files in
the correct order so that the desired effect is achieved.